don't show progress UI in private browsing
This isn't a perfect solution: not only will it not show the progress UI, it
won't show any status (even cached status) in the panel for private browsing
windows.

Going to do this for now, to avoid giving users the false impression that their
private browsing data is being sent to a 3rd party server (it is not).