checkbox controls popup notifications
Previously "enable heartbleed notifications" basically allowed you to
enable/disable the add-on. I don't see why we should provide that functionality
if you can just enable/disable the addon in about:addons.

Modified so it now controls whether popup notifications are shown on vulnerable
sties. The sites are still checked, and the status is still available in the panel.

Adds note to the first-run page explaining this (the only available setting).